requirements.txt
numpy>=1.19.0
torch>=1.7.0
opencv-python>=4.2.0
matplotlib>=3.2.0
tensorboard>=2.2.0
tqdm>=4.46.0
ale-py>=0.7.4
AutoROM.accept-rom-license
gymnasium>=0.26.0
pygame>=2.1.0
pillow>=8.0.0